requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.16.1
fsspec==2025.3.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.6
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
paho-mqtt==2.1.0
pkg_resources==0.0.0
pycryptodome==3.23.0
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
